update ivy translations to use classifiers
this is a dirty, dirty hack! but it lets us download dependency native
jars for libraries like LWJGL and put them on the classpath correctly

Ivy seems to accept classifiers for single-artifact resolution,
but not whole-project dependency resolution. It gets confused and can't
find the pom file anymore.

Ivy was just not designed to handle multiple artifacts per dependency =(